memory subsystem refactor - allow fragmentation if frag can contain a free block - start to remove pool->quanta - remove platform specific alignment - remove arena merge (not available when using malloc) - shrink paddings to the minimum

+++ include/pool.h
@@ -6,7 +6,7 @@
 extern	void	(*poolfault)(void *, char *);
 extern	void	poolinit(void);
 extern	ulong	poolmax(Pool*);
-extern	void*	dopoolalloc(Pool*, ulong);
+extern	void*	dopoolalloc(Pool*, size_t);
 extern	void*	poolalloc(Pool*, ulong);
 extern	void	poolfree(Pool*, void*);
 extern	Bhdr*	poolchain(Pool*);
